pyproject.toml
[project]
name = "bmhe_analytics_backend"
version = "0.1.0"
description = "Backend do projeto MCP - com FastAPI, segurança e autenticação"
authors = [{ name = "Seu Nome", email = "seu@email.com" }]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy>=2.0",
    "alembic",
    "pydantic>=2.0",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-dotenv",
    "httpx",
    "requests",
    "email-validator",
    "twilio"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "ipython"
]

[build-system]
requires = ["uv"]
build-backend = "uv.build"
